Merge of new position code The new changes for v2.4 (see first post in the thread) have been added to the modular build code.
Rewrite of FNC_GetPlayerUID
I have also rewritten the FNC_getplayerUID function whic was added to Epoch 1.0.5.2 (now 1.0.6) to allow server owners to switch between using the SteamID and tyhe old BIS PUID.  As it was never officially released and as the vast majority of people will now be using the SteamID I have modified it to check if the A Plot for Life variable is set and to return either the SteamID or CharacterID.  This means instead of 5 lines of code in each of the building sqf files I can now put only one or two (dep if the worldspace needs to be set).

New function FNC_GetPoles
I have also added a new function called fn_get_poles.sqf.  Quite a few of the building scripts need to check if a plot pole is nearby and if so get the nearest.  The new function does this and returns the count of plot poles and the nearest alive pole.  Using it reduces the lines of sqf code by about 2 for each file it is called in.